+++ indico/modules/admin/views.py
@@ -16,11 +16,12 @@ class WPAdmin(WPJinjaMixin, WPDecorated):
     """Base class for admin pages."""
 
     def __init__(self, rh, active_menu_item=None, **kwargs):
-        kwargs['active_menu_item'] = active_menu_item or self.sidemenu_option
+        self.active_menu_item = active_menu_item or self.sidemenu_option
+        kwargs['active_menu_item'] = self.active_menu_item
         WPDecorated.__init__(self, rh, **kwargs)
 
     def _get_breadcrumbs(self):
-        menu_item = get_menu_item('admin-sidemenu', self._kwargs['active_menu_item'])
+        menu_item = get_menu_item('admin-sidemenu', self.active_menu_item)
         items = [Breadcrumb(_('Administration'), url_for('core.admin_dashboard'))]
         if menu_item:
             items.append(menu_item.title)
@@ -28,3 +29,10 @@ class WPAdmin(WPJinjaMixin, WPDecorated):
 
     def _get_body(self, params):
         return self._get_page_content(params)
+
+    @property
+    def _extra_title_parts(self):
+        title_parts = super()._extra_title_parts
+        if menu_item := get_menu_item('admin-sidemenu', self.active_menu_item):
+            return (*title_parts, menu_item.title)
+        return title_parts